requirements.txt
uptime
python-telegram-bot
psutil
requests